Make sure gomod-k8s-dependencies bumps provide a version
in the commit message and pr title.

Since there are usually multiple dependencies updates as part of the
group, one name is used for the title.

I think it is better to have one dependency name in the title with the
according version then to just have gomod-k8s-dependencies updated to
some version.

To prevent titles like "Update gomod-k8s-dependencies (main)".